fixed output adaptor

+++ miners/adaptors/output_adaptors/spotltl_to_spotltl/run.sh
@@ -12,4 +12,9 @@ if [ ! -f $in_path ]; then
     exit 1
 fi
 
+if [ -f "$out_path" ]; then
+  echo "Error: Output path '$out_path' is a file, not a directory."
+  exit 1
+fi
+
 cp -r $in_path $out_path
